backend/requirements.txt
tensorflow
keras
numpy
pillow
scikit-image
matplotlib
uvicorn
python-dotenv
groq
fastapi
opencv-python
pydantic
python-multipart
onnxruntime
memory-profiler